Fix unreliable tmpfs test and add by-size
Using a regex to match a subdirectory of a parent tmpfs might fail.
This switches to using stat instead of df so that we can pass any
arbitrary path and see if it is a tmpfs.

by-size allow returning the first matched largest tmpfs path.

Always look for path ending in user's UID.